refactor: move Token to token package
