+++ src/engine/src/rabbitescape/engine/behaviours/Bashing.java
@@ -2,7 +2,7 @@ package rabbitescape.engine.behaviours;
 
 import static rabbitescape.engine.ChangeDescription.State.*;
 import static rabbitescape.engine.Direction.*;
-import static rabbitescape.engine.Token.Type.*;
+import static rabbitescape.engine.token.Token.Type.*;
 
 import java.util.Map;
