+++ src/engine/src/rabbitescape/engine/behaviours/Brollychuting.java
@@ -1,7 +1,7 @@
 package rabbitescape.engine.behaviours;
 
 import static rabbitescape.engine.ChangeDescription.State.*;
-import static rabbitescape.engine.Token.Type.brolly;
+import static rabbitescape.engine.token.Token.Type.brolly;
 
 import java.util.Map;
