+++ src/engine/src/rabbitescape/engine/solution/SandboxGame.java
@@ -12,7 +12,7 @@ import rabbitescape.engine.IgnoreWorldStatsListener;
 import rabbitescape.engine.Pipe;
 import rabbitescape.engine.Rabbit;
 import rabbitescape.engine.Thing;
-import rabbitescape.engine.Token;
+import rabbitescape.engine.token.Token;
 import rabbitescape.engine.VoidMarkerStyle;
 import rabbitescape.engine.World;
 import rabbitescape.engine.textworld.Comment;
