+++ src/engine/test/rabbitescape/engine/solution/TestSolution.java
@@ -5,7 +5,7 @@ import static org.hamcrest.MatcherAssert.*;
 
 import org.junit.*;
 
-import rabbitescape.engine.Token;
+import rabbitescape.engine.token.Token;
 import rabbitescape.engine.World;
 
 public class TestSolution
