+++ src/engine/test/rabbitescape/engine/solution/TestSolutionParser.java
@@ -5,7 +5,7 @@ import org.hamcrest.Description;
 import org.hamcrest.Matcher;
 import org.junit.*;
 
-import rabbitescape.engine.Token;
+import rabbitescape.engine.token.Token;
 import rabbitescape.engine.World;
 
 import static org.junit.Assert.fail;
