+++ src/render/rabbit-escape-render.iml
@@ -12,5 +12,6 @@
     <orderEntry type="sourceFolder" forTests="false" />
     <orderEntry type="module" module-name="src/engine" />
     <orderEntry type="library" scope="TEST" name="lib" level="project" />
+    <orderEntry type="module" module-name="rabbit-escape-engine" />
   </component>
 </module>
