+++ src/ui-text/rabbit-escape-ui-text.iml
@@ -13,5 +13,7 @@
     <orderEntry type="library" name="lib" level="project" />
     <orderEntry type="module" module-name="src/render" />
     <orderEntry type="module" module-name="src/engine" />
+    <orderEntry type="module" module-name="rabbit-escape-engine" />
+    <orderEntry type="module" module-name="rabbit-escape-render" />
   </component>
 </module>
